* commit '1ced6b71b3731d9907735bf7cac57f5c4b0813b4':
  VFS-5697 pr tasks done
  VFS-5697 added removing old js-client files
  VFS-5697 fixed wrong js-client branch name in git detached state

+++ Makefile
@@ -11,7 +11,11 @@ SWAGGER_JS_CLIENT_IMAGE     ?= docker.onedata.org/swagger-codegen:VFS-3144
 SWAGGER_BASH_CLIENT_IMAGE   ?= docker.onedata.org/swagger-codegen:ID-2fc8126ac8
 SWAGGER_REDOC_IMAGE         ?= docker.onedata.org/swagger-redoc:1.0.0
 
-BRANCH = $(shell git rev-parse --abbrev-ref HEAD)
+ifndef BRANCH
+	# BRANCH should always be passed in case of git detached state. This is only a fallback.
+	BRANCH = $(shell git rev-parse --abbrev-ref HEAD)
+endif
+
 COMMIT_MESSAGE = $(shell git log -1 --pretty=format:%s)
 
 .PHONY : all swagger.json
@@ -46,10 +50,12 @@ javascript-client: validate
 	docker run --rm -e "CHOWNUID=${UID}" -v `pwd`:/swagger -t ${SWAGGER_JS_CLIENT_IMAGE}  generate -i ./swagger.json -l javascript -o ./generated/javascript/ -c ./javascript-config.json
 
 javascript-update-repo: clean javascript-client
+	if [ "${BRANCH}" = "HEAD" ]; then exit 1; fi
 	rm -rf generated/javascript-git
 	git clone ssh://git@git.onedata.org:7999/vfs/onepanel-javascript-client.git generated/javascript-git && \
 	cd generated/javascript-git && \
 	( ( git checkout ${BRANCH} && ( git pull || exit 1 ) ) || git checkout -b ${BRANCH} ) && \
+	git rm -r * && \
 	cp -R ../javascript/* . && \
 	git add -A . && \
 	git config user.email "bamboo@onedata.org" && \
